Added audio spectrum page

+++ rx.cpp
@@ -329,7 +329,7 @@ void rx::get_spectrum(uint8_t spectrum[], uint8_t &dB10, uint8_t zoom)
   rx_dsp_inst.get_spectrum(spectrum, dB10, zoom);
 }
 
-void rx::get_audio(uint8_t audio[])
+void rx::get_audio(int16_t audio[])
 {
   rx_dsp_inst.get_audio_capture(audio);
 }
